need to compare primitive short
git-svn-id: https://svn.code.sf.net/p/jackcess/code/jackcess/trunk@928 f203690c-595d-4dc9-a70b-905162fa7fd2

+++ src/main/java/com/healthmarketscience/jackcess/impl/query/QueryImpl.java
@@ -303,7 +303,7 @@ public abstract class QueryImpl implements Query
         comboJoins = new ArrayList<Row>();
         comboJoinMap.put(key, comboJoins);
       } else {
-        if(comboJoins.get(0).flag != join.flag) {
+        if(comboJoins.get(0).flag != (short)join.flag) {
           throw new IllegalStateException(withErrorContext(
               "Mismatched join flags for combo joins"));
         }
